requirements.txt
playwright==1.45.1
pytest-playwright==0.5.1
python-dotenv==1.0.1